refactor: share mobile wallet factory wiring
