+++ waltid-libraries/protocols/waltid-openid4vc-wallet-mobile/src/iosMain/kotlin/id/walt/wallet2/mobile/MobileWalletFactory.kt
@@ -3,30 +3,12 @@ package id.walt.wallet2.mobile
 import id.walt.wallet2.persistence.db.WalletPersistenceDatabase
 import id.walt.wallet2.persistence.keys.IosPlatformKeyProvider
 import id.walt.wallet2.persistence.stores.DriverFactory
-import id.walt.wallet2.persistence.stores.PlatformKeyStore
-import id.walt.wallet2.persistence.stores.SqlDelightCredentialStore
-import id.walt.wallet2.persistence.stores.SqlDelightDidStore
 
 actual class MobileWalletFactory {
     actual fun create(config: MobileWalletConfig): MobileWallet {
         val driver = DriverFactory().createDriver("wallet_${config.walletId}")
         val db = WalletPersistenceDatabase(driver)
-        val queries = db.walletPersistenceQueries
-
         val keyProvider = IosPlatformKeyProvider()
-        val keyStore = PlatformKeyStore(keyProvider, queries)
-        val credentialStore = SqlDelightCredentialStore(queries)
-        val didStore = SqlDelightDidStore(queries)
-
-        return MobileWallet(
-            walletId = config.walletId,
-            keyStore = keyStore,
-            didStore = didStore,
-            credentialStore = credentialStore,
-            keyGenerator = { keyType -> keyProvider.generateKey(keyType) },
-            defaultKeyType = config.defaultKeyType,
-            attestationConfig = config.attestationConfig,
-            onEvent = config.onEvent,
-        )
+        return createMobileWallet(config, db, keyProvider)
     }
 }
