+++ waltid-libraries/protocols/waltid-openid4vc-wallet-persistence-mobile/src/iosMain/kotlin/id/walt/wallet2/persistence/keys/IosPlatformKeyProvider.kt
@@ -10,11 +10,7 @@ class IosPlatformKeyProvider(
 ) : PlatformKeyProvider {
 
     override val supportedPlatformKeyTypes: Set<KeyType> =
-        setOf(KeyType.secp256r1, KeyType.secp384r1, KeyType.secp521r1, KeyType.RSA)
-
-    override val isPlatformBackingAvailable: Boolean = true
-
-    override fun isPlatformBacked(keyType: KeyType): Boolean = keyType in supportedPlatformKeyTypes
+        PlatformKeyProvider.DEFAULT_SUPPORTED_PLATFORM_KEY_TYPES
 
     override suspend fun generateKey(keyType: KeyType, keyId: String?): Key {
         val kid = keyId ?: Uuid.random().toString()
